% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_potential.R
\name{aggregate_potential}
\alias{aggregate_potential}
\title{aggregate_potential}
\usage{
aggregate_potential(aoi_sf, pnt_sf, id_col, crs = 4647)
}
\arguments{
\item{aoi_sf}{sf of geometry type polygon used for aggregation}

\item{pnt_sf}{sf of geometry type point including attribute to be aggregated}

\item{id_col}{name of the id column used for grouping operations}

\item{crs}{epsg code}
}
\value{
sf of geometry type polygon including aggregated weighted
target potentials
}
\description{
Aggregate the sf (geometry type point) attributes
    \code{"MeanPot"} and \code{"MedianPot"} to an area of interest
}
\examples{
data(testdata)
pnt_sf <- create_potential_sf(odm_object = testdata$od_result,
                              filter_value_type = "distance",
                              accessibility_filter_value = 750,
                              search_direction = "to_target",
                              ors_profile = "foot-walking")
output_sf <- aggregate_potential(aoi_sf = testdata$grid_500m,
                                 pnt_sf = pnt_sf,
                                 id_col = "id",
                                 crs = 4647)
}
\seealso{
\code{create_potential_sf()} to create a suitable sf
}
