% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_driving_car_decay_distance.R
\name{apply_driving_car_decay_distance}
\alias{apply_driving_car_decay_distance}
\title{apply_driving_car_decay_distance}
\usage{
apply_driving_car_decay_distance(distance_value)
}
\arguments{
\item{distance_value}{A numeric value: meters}
}
\value{
A numeric value
}
\description{
apply_driving_car_decay_distance
}
