% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_foot_walking_decay_time.R
\name{apply_foot_walking_decay_time}
\alias{apply_foot_walking_decay_time}
\title{apply_foot_walking_decay_time}
\usage{
apply_foot_walking_decay_time(time_value)
}
\arguments{
\item{time_value}{A numeric value: minutes}
}
\value{
A numeric value
}
\description{
apply_foot_walking_decay_time
}
