% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_threshold.R
\name{apply_threshold}
\alias{apply_threshold}
\title{apply_threshold}
\usage{
apply_threshold(pnt_sf, threshold_col, value)
}
\arguments{
\item{pnt_sf}{sf of geometry type point including attribute to be aggregated}

\item{threshold_col}{Name of the threshold value column}

\item{value}{A numeric threshold value}
}
\value{
sf of geometry type point
}
\description{
apply_threshold
}
