% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_n_min_values.R
\name{average_n_min_values}
\alias{average_n_min_values}
\title{average_n_min_values}
\usage{
average_n_min_values(min_values_df, n_vals)
}
\arguments{
\item{min_values_df}{A tibble with n minimum distances or durations for each
feature}

\item{n_vals}{An integer defining the number of minimum values to be
incorporated}
}
\value{
A tibble including a new column of average distances or durations for
each feature
}
\description{
average_n_min_values
}
