% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_chunk.R
\name{build_chunk}
\alias{build_chunk}
\title{build_chunk}
\usage{
build_chunk(coord_list, max_chunk_size)
}
\arguments{
\item{coord_list}{A list of x-y-coordinates}

\item{max_chunk_size}{An integer value defining the maximum chunk size}
}
\value{
A list
}
\description{
build_chunk
}
