% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_n_sf.R
\name{build_n_sf}
\alias{build_n_sf}
\title{build_n_sf}
\usage{
build_n_sf(source_sf, nearest_target_dist, nearest_target_time, target_df, n)
}
\arguments{
\item{source_sf}{sf object of geometry type point containing the sources
with coordinates in separate x, y columns}

\item{nearest_target_dist}{A tibble with n minimum distances and the average
distance for each feature}

\item{nearest_target_time}{A tibble with n minimum durations and the average
duration for each feature}

\item{target_df}{data.frame containing the targets with coordinates
in separate x, y columns}

\item{n}{An integer defining the number of minimum values to be
incorporated}
}
\value{
sf object of geometry type point
}
\description{
build_n_sf
}
