% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_sf.R
\name{build_sf}
\alias{build_sf}
\title{build_sf}
\usage{
build_sf(source_sf, nearest_target_dist, nearest_target_time, target_df)
}
\arguments{
\item{source_sf}{sf object of geometry type point containing the sources
with coordinates in separate x, y columns}

\item{nearest_target_dist}{A tibble with the minimum distance}

\item{nearest_target_time}{A tibble with the minimum duration}

\item{target_df}{data.frame containing the targets with coordinates
in separate x, y columns}
}
\value{
sf object of geometry type point
}
\description{
build_sf
}
