% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_aoi_accumulation.R
\name{calc_aoi_accumulation}
\alias{calc_aoi_accumulation}
\title{calc_aoi_accumulation}
\usage{
calc_aoi_accumulation(aoi_sf, pnt_sf)
}
\arguments{
\item{aoi_sf}{sf of geometry type polygon used for aggregation}

\item{pnt_sf}{sf of geometry type point including attribute to be aggregated}
}
\value{
sf of geometry type polygon including aggregated target counts
}
\description{
calc_aoi_accumulation
}
