% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_aoi_coverage_rate.R
\name{calc_aoi_coverage_rate}
\alias{calc_aoi_coverage_rate}
\title{calc_aoi_coverage_rate}
\usage{
calc_aoi_coverage_rate(aoi_sf, pnt_sf, id_col)
}
\arguments{
\item{aoi_sf}{sf of geometry type polygon used for aggregation}

\item{pnt_sf}{sf of geometry type point including attribute to be aggregated}

\item{id_col}{name of the id column used for grouping operations}
}
\value{
sf of geometry type polygon including an aggregated coverage rate
target potentials
}
\description{
calc_aoi_coverage_rate
}
