% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_aoi_expenses.R
\name{calc_aoi_expenses}
\alias{calc_aoi_expenses}
\title{calc_aoi_expenses}
\usage{
calc_aoi_expenses(aoi_sf, pnt_sf)
}
\arguments{
\item{aoi_sf}{sf of geometry type polygon used for aggregation}

\item{pnt_sf}{sf of geometry type point including attribute to be aggregated}
}
\value{
sf of geometry type polygon including aggregated distance and time
values
}
\description{
calc_aoi_expenses
}
