% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_aoi_potential.R
\name{calc_aoi_potential}
\alias{calc_aoi_potential}
\title{calc_aoi_potential}
\usage{
calc_aoi_potential(aoi_sf, pnt_sf)
}
\arguments{
\item{aoi_sf}{sf of geometry type polygon used for aggregation}

\item{pnt_sf}{sf of geometry type point including attribute to be aggregated}
}
\value{
sf of geometry type polygon including aggregated weighted target
potentials
}
\description{
calc_aoi_potential
}
