% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_ors_matrix.R
\name{call_ors_matrix}
\alias{call_ors_matrix}
\title{call_ors_matrix}
\usage{
call_ors_matrix(
  x_y_coordinates,
  ors_profile,
  sources_chunk_size,
  destination_chunk_size
)
}
\arguments{
\item{x_y_coordinates}{List of longitude, latitude coordinate pairs}

\item{ors_profile}{Route profile}

\item{sources_chunk_size}{integer value to identify the sources}

\item{destination_chunk_size}{integer value to identify the targets}
}
\value{
Duration and distance matrix for multiple sources and targets
}
\description{
Chunk-wise calls to ors_matrix
}
