% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_targets.R
\name{count_targets}
\alias{count_targets}
\title{count_targets}
\usage{
count_targets(data)
}
\arguments{
\item{data}{A tibble}
}
\value{
A tibble with a count of targets for each id
}
\description{
count_targets
}
