% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cumulative_by_cat_sf.R
\name{create_cumulative_by_cat_sf}
\alias{create_cumulative_by_cat_sf}
\title{create_cumulative_by_cat_sf}
\usage{
create_cumulative_by_cat_sf(
  odm_object,
  source_sf,
  target_df,
  filter_value_type,
  accessibility_filter_value,
  search_direction,
  filter_attribute,
  filter_value
)
}
\arguments{
\item{odm_object}{of type list encompassing duration and distance
matrices and the source and target locations}

\item{source_sf}{sf object of geometry type point containing the sources
with coordinates in separate x, y columns}

\item{target_df}{data.frame containing the targets with coordinates
in separate x, y columns}

\item{filter_value_type}{A string, "time" or "distance"}

\item{accessibility_filter_value}{A numeric value of minutes or meters}

\item{search_direction}{A string, "to_targets" or "from_targets"}

\item{filter_attribute}{A category column to subset the targets}

\item{filter_value}{A string with a specific category value}
}
\value{
sf object of geometry type point
}
\description{
Creates sf object of geometry type point with a count of sources
for each target id for all features matching the applied filters
}
