% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_potential_by_cat_sf.R
\name{create_potential_by_cat_sf}
\alias{create_potential_by_cat_sf}
\title{create_potential_by_cat_sf}
\usage{
create_potential_by_cat_sf(
  odm_object,
  filter_value_type,
  accessibility_filter_value,
  search_direction,
  filter_attribute,
  filter_value,
  ors_profile
)
}
\arguments{
\item{odm_object}{of type list encompassing a duration and a distance
matrix, the source and target locations. Created by using either
\code{calc_odm} or \code{calc_odm_multicore}}

\item{filter_value_type}{A string, "time" or "distance"}

\item{accessibility_filter_value}{A numeric value of minutes or meters}

\item{search_direction}{A string, "to_targets" or "from_targets"}

\item{filter_attribute}{A category column to subset the targets}

\item{filter_value}{A string with a specific category value}

\item{ors_profile}{Route profile, e.g. \code{"foot-walking"} or
\code{"driving-car"}}
}
\value{
sf object of geometry type point
}
\description{
Creates sf of geometry type point with a weighted value
    of features belonging to a specific category that match the applied time
    or distance filter
}
\examples{
data(testdata)
od_result <- testdata$od_result
output_sf <- create_potential_by_cat_sf(odm_object = od_result,
                                        filter_value_type = "distance",
                                        accessibility_filter_value = 750,
                                        search_direction = "to_target",
                                        filter_attribute = "category",
                                        filter_value = "A",
                                        ors_profile = "foot-walking")
}
\seealso{
\code{calc_odm()} or \code{calc_odm_multicore()} for odm_object
    creation and \code{create_potential_sf()} to create a weighted value
    for all features up to a threshold distance or time value
}
