% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_potential_sf.R
\name{create_potential_sf}
\alias{create_potential_sf}
\title{create_potential_sf}
\usage{
create_potential_sf(
  odm_object,
  source_sf,
  target_df,
  filter_value_type,
  accessibility_filter_value,
  search_direction = "toTarget",
  ors_profile
)
}
\arguments{
\item{odm_object}{of type list encompassing duration and distance
matrices and the source and target locations}

\item{source_sf}{sf object of geometry type point containing the sources
with coordinates in separate x, y columns}

\item{target_df}{data.frame containing the targets with coordinates
in separate x, y columns}

\item{filter_value_type}{A string, "time" or "distance"}

\item{accessibility_filter_value}{A numeric value of minutes or meters}

\item{search_direction}{A string, "to_targets" or "from_targets"}

\item{ors_profile}{Route profile}
}
\value{
sf object of geometry type point
}
\description{
Creates sf object of geometry type point with a weighted value
of sources for each target id for all features matching the applied filters
}
