% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_time_distance_by_cat_sf.R
\name{create_time_distance_by_cat_sf}
\alias{create_time_distance_by_cat_sf}
\title{create_time_distance_by_cat_sf}
\usage{
create_time_distance_by_cat_sf(odm_object, filter_attribute, filter_value)
}
\arguments{
\item{odm_object}{of type list encompassing a duration and a distance
matrix, the source and target locations. Created by using either
\code{calc_odm} or \code{calc_odm_multicore}}

\item{filter_attribute}{A category column to subset the targets}

\item{filter_value}{A string with a specific category value}
}
\value{
sf object of geometry type point
}
\description{
Creates sf of geometry type point for each source feature with
    the minimum distance and duration values to the nearest target feature
    which matches the applied category filter
}
\examples{
data(testdata)
od_result <- testdata$od_result
output_sf <- create_time_distance_by_cat_sf(odm_object = od_result,
    filter_attribute = "category", filter_value = "A")
}
\seealso{
\code{calc_odm()} or \code{calc_odm_multicore()} for odm_object
    creation and \code{create_time_distance_n_by_ca_sf} to calculate the n
    nearest target locations for a specific category
}
