% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_time_distance_n_sf.R
\name{create_time_distance_n_sf}
\alias{create_time_distance_n_sf}
\title{Calculate n closest distances and durations}
\usage{
create_time_distance_n_sf(odm_object, n)
}
\arguments{
\item{odm_object}{of type list encompassing a duration and a distance
matrix, the source and target locations. Created by using either
\code{calc_odm} or \code{calc_odm_multicore}}

\item{n}{An integer defining the number of minimum values to be
incorporated}
}
\value{
sf object of geometry type point
}
\description{
Creates sf of geometry type point for each source feature with
    the minimum distance and duration values to n nearest target features.
}
\examples{
data(testdata)
od_result <- testdata$od_result
# Use e.g. n = 3
output_sf <- create_time_distance_n_sf(odm_object = od_result, n = 3)
}
\seealso{
\code{calc_odm()} or \code{calc_odm_multicore()} for odm_object
    creation and \code{create_time_distance_sf()} to calculate only the
    nearest target location
}
