% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_by_distance_cat_cum.R
\name{filter_by_distance_cat_cum}
\alias{filter_by_distance_cat_cum}
\title{filter_by_distance_cat_cum}
\usage{
filter_by_distance_cat_cum(
  odm_object,
  distance_value,
  search_direction_str,
  target_df,
  filter_attribute,
  filter_value
)
}
\arguments{
\item{odm_object}{odm-object of type list encompassing duration and distance
matrices and the source and target locations}

\item{distance_value}{A numeric value of meters}

\item{target_df}{data.frame containing the targets with coordinates
in separate x, y columns}

\item{filter_attribute}{A category column to subset the targets}

\item{filter_value}{A string with a specific category value}

\item{search_direction}{A character string, "to_target" or "from_target"}
}
\value{
A tibble with a count of sources for each target id for all features
matching the applied filters
}
\description{
filter_by_distance_cat_cum
}
