% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_by_distance_cum.R
\name{filter_by_distance_cum}
\alias{filter_by_distance_cum}
\title{filter_by_distance_cum}
\usage{
filter_by_distance_cum(odm_object, distance_value, search_direction)
}
\arguments{
\item{odm_object}{odm-object of type list encompassing duration and distance
matrices and the source and target locations}

\item{search_direction}{A character string, "to_target" or "from_target"}

\item{time_value}{A numeric value of meters}
}
\value{
A tibble with a count of sources for each target id or
a count of targets for each source id
}
\description{
filter_by_distance_cum
}
