% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_by_distance_pot.R
\name{filter_by_distance_pot}
\alias{filter_by_distance_pot}
\title{filter_by_distance_pot}
\usage{
filter_by_distance_pot(
  odm_object,
  distance_value,
  search_direction,
  ors_profile
)
}
\arguments{
\item{odm_object}{odm-object of type list encompassing duration and distance
matrices and the source and target locations}

\item{distance_value}{A numeric value of meters}

\item{search_direction}{A character string, "to_target" or "from_target"}

\item{ors_profile}{Route profile}
}
\value{
A tibble with a weighted value of sources for each target id or
a weighted value of targets for each source id
}
\description{
filter_by_distance_pot
}
