% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_by_time_cum.R
\name{filter_by_time_cum}
\alias{filter_by_time_cum}
\title{filter_by_time_cum}
\usage{
filter_by_time_cum(odm_object, time_value, search_direction)
}
\arguments{
\item{odm_object}{odm-object of type list encompassing duration and distance
matrices and the source and target locations}

\item{time_value}{A numeric value of minutes}

\item{search_direction}{A character string, "to_target" or "from_target"}
}
\value{
A tibble with a count of sources for each target id or
a count of targets for each source id
}
\description{
filter_by_time_cum
}
