% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_by_time_pot.R
\name{filter_by_time_pot}
\alias{filter_by_time_pot}
\title{filter_by_time_pot}
\usage{
filter_by_time_pot(odm_object, time_value, search_direction, ors_profile)
}
\arguments{
\item{odm_object}{odm-object of type list encompassing duration and distance
matrices and the source and target locations}

\item{time_value}{A numeric value of minutes}

\item{search_direction}{A character string, "to_target" or "from_target"}

\item{ors_profile}{Route profile}
}
\value{
A tibble with a weighted value of sources for each target id or
a weighted value of targets for each source id
}
\description{
filter_by_time_pot
}
