% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_odm.R
\name{filter_odm}
\alias{filter_odm}
\title{filter_odm}
\usage{
filter_odm(matrix_data, target_df, filter_attribute, filter_value)
}
\arguments{
\item{matrix_data}{Distance or duration matrix}

\item{target_df}{data.frame containing the sources with coordinates
in separate x, y columns}

\item{filter_attribute}{A category column to subset the targets}

\item{filter_value}{A string with a specific category value}
}
\value{
A matrix including only features matching the applied filter
}
\description{
filter_odm
}
