% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_cum_out_from_target_sf.R
\name{format_cum_out_from_target_sf}
\alias{format_cum_out_from_target_sf}
\title{format_cum_out_from_target_sf}
\usage{
format_cum_out_from_target_sf(target_df, target_cnt_df)
}
\arguments{
\item{target_cnt_df}{A tibble with a count of targets for each source id}

\item{sources_sf}{sf object of geometry type point containing the sources
with coordinates in separate x, y columns}
}
\value{
sf object of geometry type point
}
\description{
format_cum_out_from_target_sf
}
