% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_pot_out_to_target_sf.R
\name{format_pot_out_to_target_sf}
\alias{format_pot_out_to_target_sf}
\title{format_pot_out_to_target_sf}
\usage{
format_pot_out_to_target_sf(in_sf, target_pot_df)
}
\arguments{
\item{target_pot_df}{A tibble with a weighted value of targets for each
source id}

\item{sources_sf}{sf object of geometry type point containing the sources
with coordinates in separate x, y columns}
}
\value{
sf object of geometry type point
}
\description{
format_pot_out_to_target_sf
}
