% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_results.R
\name{format_results}
\alias{format_results}
\title{format_results}
\usage{
format_results(matrix_object)
}
\arguments{
\item{matrix_object}{A time or distance matrix}
}
\value{
A tibble
}
\description{
format_results
}
