% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_results_by_cat.R
\name{format_results_by_cat}
\alias{format_results_by_cat}
\title{format_results_by_cat}
\usage{
format_results_by_cat(matrix_object, target_df, filter_attribute, filter_value)
}
\arguments{
\item{matrix_object}{A time or distance matrix}

\item{target_df}{A dataframe containing the targets}

\item{filter_attribute}{A category column to subset the targets}

\item{filter_value}{A string with a specific category value}
}
\value{
A tibble
}
\description{
format_results_by_cat
}
