% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_target.R
\name{from_target}
\alias{from_target}
\title{from_target}
\usage{
from_target(out_mat, odm_object, time_distance_value)
}
\arguments{
\item{out_mat}{A duration or distance matrix object}

\item{odm_object}{odm-object of type list encompassing duration and distance
matrices and the source and target locations}

\item{time_distance_value}{A numeric value of minutes or meters defining a
threshold filter value}
}
\value{
A tibble with a count of sources for each target id
}
\description{
from_target
}
