% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_target_by_cat.R
\name{from_target_by_cat}
\alias{from_target_by_cat}
\title{from_target_by_cat}
\usage{
from_target_by_cat(
  out_mat,
  odm_object,
  time_distance_value,
  target_df,
  filter_attribute,
  filter_value
)
}
\arguments{
\item{out_mat}{A duration or distance matrix object}

\item{odm_object}{odm-object of type list encompassing duration and distance
matrices and the source and target locations}

\item{time_distance_value}{A numeric value of minutes or meters defining a
threshold filter value}

\item{target_df}{data.frame containing the targets with coordinates
in separate x, y columns}

\item{filter_attribute}{A category column to subset the targets}

\item{filter_value}{A string with a specific category value}
}
\value{
A tibble with a count of sources for each target id for all features
matching the applied filter
}
\description{
from_target_by_cat
}
