% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_target_by_cat_distance_weighted.R
\name{from_target_by_cat_distance_weighted}
\alias{from_target_by_cat_distance_weighted}
\title{from_target_by_cat_distance_weighted}
\usage{
from_target_by_cat_distance_weighted(
  out_mat,
  odm_object,
  time_distance_value,
  target_df,
  filter_attribute,
  filter_value,
  ors_profile
)
}
\arguments{
\item{out_mat}{A distance matrix object}

\item{odm_object}{of type list encompassing duration and distance
matrices and the source and target locations}

\item{time_distance_value}{A numeric value of meters}

\item{target_df}{data.frame containing the targets with coordinates
in separate x, y columns}

\item{filter_attribute}{A category column to subset the targets}

\item{filter_value}{A string with a specific category value}

\item{ors_profile}{Route profile}
}
\value{
A tibble
}
\description{
from_target_by_cat_distance_weighted
}
