% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_target_distance_weighted.R
\name{from_target_distance_weighted}
\alias{from_target_distance_weighted}
\title{from_target_distance_weighted}
\usage{
from_target_distance_weighted(
  out_mat,
  odm_object,
  time_distance_value,
  ors_profile
)
}
\arguments{
\item{out_mat}{A distance matrix object}

\item{odm_object}{of type list encompassing duration and distance
matrices and the source and target locations}

\item{time_distance_value}{A numeric value of meters}

\item{ors_profile}{Route profile}
}
\value{
A tibble
}
\description{
from_target_distance_weighted
}
