% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coords_as_list.R
\name{get_coords_as_list}
\alias{get_coords_as_list}
\title{get_coords_as_list}
\usage{
get_coords_as_list(data)
}
\arguments{
\item{data}{A data.frame with coordinates in x and y columns}
}
\value{
List of longitude, latitude coordinate pairs
}
\description{
get_coords_as_list
}
