% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_min_values.R
\name{get_min_values}
\alias{get_min_values}
\title{get_min_values}
\usage{
get_min_values(matrix_data)
}
\arguments{
\item{matrix_data}{Distance or duration matrix}
}
\value{
A tibble with the minimum distance or duration for each feature
}
\description{
get_min_values
}
