% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_n_min_values.R
\name{get_n_min_values}
\alias{get_n_min_values}
\title{get_n_min_values}
\usage{
get_n_min_values(matrix_data, n_vals)
}
\arguments{
\item{matrix_data}{Distance or duration matrix}

\item{n_vals}{An integer defining the number of minimum values to be
extracted}
}
\value{
A tibble with n minimum distances or durations for each feature
}
\description{
get_n_min_values
}
