% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_n_nearest_targets.R
\name{identify_n_nearest_targets}
\alias{identify_n_nearest_targets}
\title{identify_n_nearest_targets}
\usage{
identify_n_nearest_targets(odm_object, matrix_property, n)
}
\arguments{
\item{odm_object}{of type list encompassing duration and distance
matrices and the source and target locations}

\item{matrix_property}{A string, "distance" or "duration"}

\item{n}{An integer defining the number of minimum values to be
extracted}
}
\value{
A tibble with n minimum distances or durations for each feature and
average distances or durations for each feature
}
\description{
identify_n_nearest_targets
}
