% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_n_nearest_targets_by_cat.R
\name{identify_n_nearest_targets_by_cat}
\alias{identify_n_nearest_targets_by_cat}
\title{identify_n_nearest_targets_by_cat}
\usage{
identify_n_nearest_targets_by_cat(
  odm_object,
  matrix_property,
  target_df,
  filter_attribute,
  filter_value,
  n
)
}
\arguments{
\item{odm_object}{of type list encompassing duration and distance
matrices and the source and target locations}

\item{matrix_property}{A string, "distance" or "duration"}

\item{target_df}{data.frame containing the targets with coordinates
in separate x, y columns}

\item{filter_attribute}{A category column to subset the targets}

\item{filter_value}{A string with a specific category value}

\item{n}{An integer defining the number of minimum values to be
extracted}
}
\value{
A tibble with n minimum distances or durations for all features
matching the applied filter and average distances or durations for each
feature
}
\description{
identify_n_nearest_targets_by_cat
}
