% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_nearest_target.R
\name{identify_nearest_target}
\alias{identify_nearest_target}
\title{identify_nearest_target}
\usage{
identify_nearest_target(odm_object, matrix_property)
}
\arguments{
\item{odm_object}{odm-object of type list encompassing duration and distance
matrices and the source and target locations}

\item{matrix_property}{A string, "distance" or "duration"}
}
\value{
A tibble with the minimum distance or duration for each feature
}
\description{
identify_nearest_target
}
