% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_nearest_target_by_cat.R
\name{identify_nearest_target_by_cat}
\alias{identify_nearest_target_by_cat}
\title{identify_nearest_target_by_cat}
\usage{
identify_nearest_target_by_cat(
  odm_object,
  matrix_property,
  target_df,
  filter_attribute,
  filter_value
)
}
\arguments{
\item{matrix_property}{A string, "distance" or "duration"}

\item{target_df}{data.frame containing the targets with coordinates
in separate x, y columns}

\item{filter_attribute}{A category column to subset the targets}

\item{filter_value}{A string with a specific category value}

\item{odm-object}{of type list encompassing duration and distance
matrices and the source and target locations}
}
\value{
A tibble with the minimum distance or duration for all features
matching the applied filter
}
\description{
identify_nearest_target_by_cat
}
