% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_indicators.R
\name{merge_indicators}
\alias{merge_indicators}
\title{Merge indicators}
\usage{
merge_indicators(aoi_sf, id_col, ...)
}
\arguments{
\item{aoi_sf}{sf of geometry type polygon with matching geometries for all
indicators to be merged}

\item{id_col}{Name of the column to merge the indicators}

\item{...}{indicators to be merged}
}
\value{
sf of geometry type polygon including the values of all input
    indicators
}
\description{
Merge several indicators which have been aggregated to the same
    area of interest.
}
\examples{
\dontrun{
data(testdata)
merged_sf <- merge_indicators(
  aoi_sf = testdata$grid_500m,
  id_col = "id",
  aggregated_sf_1,
  aggregated_sf_2,
  aggregated_sf_3,
  ...
)
}
}
