% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_n.R
\name{min_n}
\alias{min_n}
\title{min_n}
\usage{
min_n(x, n = 1)
}
\arguments{
\item{x}{A row of a matrix}

\item{n}{An integer defining the number of minimum values to be extracted}
}
\value{

}
\description{
min_n
}
