% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_inputs.R
\name{read_inputs}
\alias{read_inputs}
\title{read_inputs}
\usage{
read_inputs(aoi_sf, pnt_sf, id_col, crs = 4647)
}
\arguments{
\item{aoi_sf}{sf of geometry type polygon used for aggregation}

\item{pnt_sf}{sf of geometry type point including attribute to be aggregated}

\item{id_col}{name of the id column used for grouping operations}

\item{crs}{epsg code}
}
\value{
A list object
}
\description{
read_inputs
}
