% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfc_as_cols.R
\name{sfc_as_cols}
\alias{sfc_as_cols}
\title{sfc_as_cols}
\usage{
sfc_as_cols(x, names = c("x", "y"))
}
\arguments{
\item{x}{A sf or sfc_POINT object}

\item{names}{A character vector of length two to define names for the new
x and y coordinate columns}
}
\value{
sf object including coordinates as new columns
}
\description{
Binds x-y-coordinates as cols to sf
}
