% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switch_names.R
\name{switch_names}
\alias{switch_names}
\title{switch_names}
\usage{
switch_names(out_mat, odm_object)
}
\arguments{
\item{out_mat}{A duration or distance matrix object}

\item{odm_object}{A list object containing a duration and a distance matrix
object, the sources and targets}
}
\value{
A matrix object with corresponding ids for sources and targets
}
\description{
Generate unique ids for sources and targets and move them to
the first column index position
}
