% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_target.R
\name{to_target}
\alias{to_target}
\title{to_target}
\usage{
to_target(out_mat, odm_object, time_distance_value)
}
\arguments{
\item{out_mat}{A duration or distance matrix object}

\item{odm_object}{odm-object of type list encompassing duration and distance
matrices
and the source and target locations}

\item{time_distance_value}{An integer of minutes or meters defining a
threshold filter value}
}
\value{
A tibble with a count of targets for each source id
}
\description{
to_target
}
