% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_target_time_weighted.R
\name{to_target_time_weighted}
\alias{to_target_time_weighted}
\title{to_target_time_weighted}
\usage{
to_target_time_weighted(out_mat, odm_object, time_distance_value, ors_profile)
}
\arguments{
\item{out_mat}{A duration matrix object}

\item{odm_object}{of type list encompassing duration and distance
matrices and the source and target locations}

\item{time_distance_value}{A numeric value of minutes}

\item{ors_profile}{Route profile}
}
\value{
A tibble
}
\description{
to_target_time_weighted
}
