% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_targets_by_distance.R
\name{weight_targets_by_distance}
\alias{weight_targets_by_distance}
\title{weight_targets_by_distance}
\usage{
weight_targets_by_distance(data, ors_profile)
}
\arguments{
\item{data}{A tibble}

\item{ors_profile}{Route profile}
}
\value{
A tibble with count values weighted by distance
}
\description{
weight_targets_by_distance
}
