% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_targets_by_time.R
\name{weight_targets_by_time}
\alias{weight_targets_by_time}
\title{weight_targets_by_time}
\usage{
weight_targets_by_time(data, ors_profile)
}
\arguments{
\item{data}{A tibble}

\item{ors_profile}{Route profile}
}
\value{
A tibble with count values weighted by time
}
\description{
weight_targets_by_time
}
