% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_coverage_rate_by_threshold.R
\name{aggregate_coverage_rate_by_threshold}
\alias{aggregate_coverage_rate_by_threshold}
\title{aggregate_coverage_rate_by_threshold}
\usage{
aggregate_coverage_rate_by_threshold(
  aoi_sf,
  pnt_sf,
  id_col,
  threshold_col,
  threshold_value,
  crs = 4647
)
}
\arguments{
\item{aoi_sf}{sf of geometry type polygon used for aggregation}

\item{pnt_sf}{sf of geometry type point including attribute to be aggregated}

\item{id_col}{name of the id column used for grouping operations}

\item{threshold_col}{Name of the threshold value column}

\item{threshold_value}{A numeric threshold value, minutes or meters}

\item{crs}{epsg code}
}
\value{
sf of geometry type polygon including an aggregated coverage rate
target potentials
}
\description{
aggregate_coverage_rate_by_threshold
}
