% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_coverage_rate_by_threshold.R
\name{aggregate_coverage_rate_by_threshold}
\alias{aggregate_coverage_rate_by_threshold}
\title{aggregate_coverage_rate_by_threshold}
\usage{
aggregate_coverage_rate_by_threshold(
  aoi_sf,
  pnt_sf,
  id_col,
  threshold_col,
  threshold_value,
  crs = 4647
)
}
\arguments{
\item{aoi_sf}{sf of geometry type polygon used for aggregation}

\item{pnt_sf}{sf of geometry type point including attribute to be aggregated}

\item{id_col}{name of the id column used for grouping operations}

\item{threshold_col}{Name of the threshold value column, \code{"Min_Dist"}
or \code{"Min_Time"}}

\item{threshold_value}{A numeric threshold value, minutes or meters}

\item{crs}{epsg code}
}
\value{
sf of geometry type polygon including an aggregated coverage rate
target potentials
}
\description{
Calculate an average coverage rate within an area of interest
    by applying a threshold filter to one of the sf (geometry type point)
    attributes \code{"Min_Dist"} or \code{"Min_Time"}. A source location
    is defined as covered if a target location is accessible within the
    threshold distance or time.
}
\examples{
data(testdata)
pnt_sf <- create_time_distance_sf(odm_object = testdata$od_result)
output_sf <- aggregate_coverage_rate_by_threshold(
    aoi_sf = testdata$grid_500m,
    pnt_sf = pnt_sf,
    id_col = "id",
    threshold_col = "Min_Dist",
    threshold_value = 350,
    crs = 4647)
}
\seealso{
\code{create_time_distance_sf()} to create a suitable sf
}
