% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_expenses.R
\name{aggregate_expenses}
\alias{aggregate_expenses}
\title{aggregate_expenses}
\usage{
aggregate_expenses(aoi_sf, pnt_sf, id_col, crs)
}
\arguments{
\item{aoi_sf}{sf of geometry type polygon used for aggregation}

\item{pnt_sf}{sf of geometry type point including attribute to be aggregated}

\item{id_col}{name of the id column used for grouping operations}

\item{crs}{epsg code}
}
\value{
sf of geometry type polygon including aggregated distance and time
values
}
\description{
aggregate_expenses
}
