% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cumulative_by_cat_sf.R
\name{create_cumulative_by_cat_sf}
\alias{create_cumulative_by_cat_sf}
\title{create_cumulative_by_cat_sf}
\usage{
create_cumulative_by_cat_sf(
  odm_object,
  filter_value_type,
  accessibility_filter_value,
  search_direction,
  filter_attribute,
  filter_value
)
}
\arguments{
\item{odm_object}{of type list encompassing a duration and a distance
matrix, the source and target locations. Created by using either
\code{calc_odm} or \code{calc_odm_multicore}}

\item{filter_value_type}{A string, "time" or "distance"}

\item{accessibility_filter_value}{A numeric value of minutes or meters}

\item{search_direction}{A string, "to_target" or "from_target"}

\item{filter_attribute}{A category column to subset the targets}

\item{filter_value}{A string with a specific category value}
}
\value{
sf object of geometry type point
}
\description{
Creates sf of geometry type point with a count of features
    that belong to a specific category and match the applied time or distance
    filter
}
\examples{
data(testdata)
od_result <- testdata$od_result
output_sf <- create_cumulative_by_cat_sf(
  odm_object = od_result,
  filter_value_type = "distance",
  accessibility_filter_value = 750,
  search_direction = "to_target",
  filter_attribute = "category",
  filter_value = "A"
)
}
\seealso{
\code{calc_odm()} or \code{calc_odm_multicore()} for odm_object
    creation and \code{create_cumulative_sf} to count features up to a
    threshold distance or time
}
