% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_time_distance_n_by_cat_sf.R
\name{create_time_distance_n_by_cat_sf}
\alias{create_time_distance_n_by_cat_sf}
\title{create_time_distance_n_by_cat_sf}
\usage{
create_time_distance_n_by_cat_sf(
  odm_object,
  source_sf,
  target_df,
  filter_attribute,
  filter_value,
  n
)
}
\arguments{
\item{odm_object}{of type list encompassing duration and distance
matrices and the source and target locations}

\item{source_sf}{sf object of geometry type point containing the sources
with coordinates in separate x, y columns}

\item{target_df}{data.frame containing the targets with coordinates
in separate x, y columns}

\item{filter_attribute}{A category column to subset the targets}

\item{filter_value}{A string with a specific category value}

\item{n}{An integer defining the number of minimum values to be
incorporated}
}
\value{
sf object of geometry type point
}
\description{
Create sf object of geometry type point with n minimum distance
and duration values for each source to target combination that matches the
applied category filter
}
