% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_time_distance_sf.R
\name{create_time_distance_sf}
\alias{create_time_distance_sf}
\title{Calculate closest distance and duration}
\usage{
create_time_distance_sf(odm_object)
}
\arguments{
\item{odm_object}{of type list encompassing a duration and a distance
matrix, the source and target locations. Created by using either
\code{calc_odm} or \code{calc_odm_multicore}}
}
\value{
sf object of geometry type point
}
\description{
Creates sf of geometry type point for each source feature with
    the minimum distance and duration values to the nearest target feature.
}
\examples{
data(testdata)
od_result <- testdata$od_result
output_sf <- create_time_distance_sf(odm_object = od_result)
}
\seealso{
\code{calc_odm()} or \code{calc_odm_multicore()} for odm_object
    creation and \code{create_time_distance_n_sf} to calculate the n nearest
    target locations
}
