% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_indicators.R
\name{merge_indicators}
\alias{merge_indicators}
\title{merge_indicators}
\usage{
merge_indicators(aoi_sf, id_col, ...)
}
\arguments{
\item{aoi_sf}{sf of geometry type polygon with matching geometries for all
indicators to be merged}

\item{id_col}{Name of the column to merge the indicators}

\item{...}{indicators to be merged}
}
\value{
sf of geometry type polygon including the values of all input
indicators
}
\description{
merge_indicators
}
